Added loading files as streams from bsp's pak files. Made models now compare their mdl, vvd, and vtx checksums to ensure they are the same before parsing.

+++ SourceModel/MDL/Structs/studiohdr_t.cs
@@ -6,7 +6,7 @@ namespace UnitySourceEngine
     {
         public int id; // Model format ID, such as "IDST" (0x49 0x44 0x53 0x54)
         public int version; // Format version number, such as 48 (0x30,0x00,0x00,0x00)
-        public long checkSum; // this has to be the same in the phy and vtx files to load!
+        public int checkSum; // this has to be the same in the phy and vtx files to load!
         public char[] name; // The internal name of the model, padding with null bytes.
                             // Typically "my_model.mdl" will have an internal name of "my_model"
         public int dataLength;  // Data size of MDL file in bytes.
